test: refactor to allow injection of mock http client on setup
