+++ packages-private/aio-commerce-lib-api/test/setup.ts
@@ -1,23 +1,25 @@
 import ky, { type KyInstance } from "ky";
 import { afterEach, beforeEach, vi } from "vitest";
 
-import { HttpClientBase } from "#lib/http-client-base";
+import type { HttpClientBase } from "#lib/http-client-base";
 
-export type TestHttpClientConfig = {
-  apiKey: string;
+type TestConfigBase = {
   baseUrl: string;
+  [key: string]: unknown;
 };
 
-// Now create your test class that inherits from the mocked base
-export class TestHttpClient extends HttpClientBase<TestHttpClientConfig> {
-  // biome-ignore lint/complexity/noUselessConstructor: False positive, as we need to override the inherited `protected` access specifier.
-  public constructor(config: TestHttpClientConfig, kyInstance: KyInstance) {
-    super(config, kyInstance);
-  }
-}
+// A constructor of a class that extends HttpClientBase.
+type TestClientFactory<
+  TConfig extends TestConfigBase,
+  TClient extends HttpClientBase<TConfig>,
+> = new (config: TConfig, kyInstance: KyInstance) => TClient;
 
 /** Performs the test setup for the library API. */
-export function libApiTestSetup() {
+export function libApiTestSetup<
+  TConfig extends TestConfigBase,
+  TClient extends HttpClientBase<TConfig>,
+  TClientFactory extends TestClientFactory<TConfig, TClient>,
+>(ClientFactory: TClientFactory, clientConfig: TConfig) {
   // A mock of the fetch function we're going to use to test the HTTP client.
   const fetch = vi.fn(async (_input, _init) => {
     return await Promise.resolve(
@@ -28,19 +30,14 @@ export function libApiTestSetup() {
     );
   });
 
-  let testClient: TestHttpClient;
-  const clientConfig: TestHttpClientConfig = {
-    apiKey: "test-api-key",
-    baseUrl: "https://api.example.com",
-  };
-
+  let testClient: TClient;
   beforeEach(() => {
     const kyClient = ky.create({
       prefixUrl: clientConfig.baseUrl,
       fetch: fetch as unknown as typeof globalThis.fetch,
     });
 
-    testClient = new TestHttpClient(clientConfig, kyClient);
+    testClient = new ClientFactory(clientConfig, kyClient);
   });
 
   afterEach(() => {
